chore: share the Tauri mesh-op staging and read primitives
meshHollowing, meshPunching, meshOrganicCut and meshRepair each carried
their own copy of the TauriInvoke type, the cached loadTauriCore
loader, the triangle-soup expander, the staging upload and the f32
payload decoder.

All four now import them from utils/tauriMeshBridge. The decoder was
spelled three ways (readStagedPositions with the command hardcoded,
readPositionsFromCommand with a per-file command union, and decodeF32
taking a label) but had one body; it is now readPositionsFromCommand
over a shared MeshPositionReadCommand union, with decodeF32 exported
for the organic-cut part reader that indexes its own label.

meshRepair's expander differed only in brace style and a loop
increment; behaviour is unchanged.
