+++ src/utils/meshRepair.ts
@@ -16,8 +16,12 @@
  * In the browser (non-Tauri) this module is inert — call sites fall back to
  * the WASM Manifold repair path.
  */
-import { isTauriRuntime } from './tauriRuntime';
 import * as THREE from 'three';
+import {
+  expandGeometryToTriangleSoup,
+  loadTauriCore,
+  readPositionsFromCommand,
+} from './tauriMeshBridge';
 
 export interface MeshAnalysisJson {
   triangle_count: number;
@@ -137,14 +141,6 @@ interface RawMeshHealthReport extends UnknownRecord {
   total_ms?: unknown;
 }
 
-type TauriInvoke = <T>(cmd: string, args?: Record<string, unknown> | ArrayBuffer | ArrayBufferView, opts?: { headers?: Record<string, string> }) => Promise<T>;
-
-interface TauriCoreModule {
-  invoke: TauriInvoke;
-}
-
-let tauriCorePromise: Promise<TauriCoreModule | null> | null = null;
-
 function asRecord(value: unknown): UnknownRecord {
   return value != null && typeof value === 'object' ? value as UnknownRecord : {};
 }
@@ -227,40 +223,6 @@ function normalizeMeshHealthReport(input: unknown): MeshHealthReport {
   };
 }
 
-
-async function loadTauriCore(): Promise<TauriCoreModule | null> {
-  if (!isTauriRuntime()) return null;
-  if (!tauriCorePromise) {
-    tauriCorePromise = import('@tauri-apps/api/core')
-      .then((mod) => ({ invoke: mod.invoke as TauriInvoke }))
-      .catch(() => null);
-  }
-  return tauriCorePromise;
-}
-
-/**
- * Reads the repaired positions from the Rust staging buffer. Tauri v2 returns
- * raw `Response` body as an ArrayBuffer automatically when the command uses
- * `tauri::ipc::Response::new(bytes)`.
- */
-async function readStagedPositions(invoke: TauriInvoke): Promise<Float32Array> {
-  const bytes = await invoke<ArrayBuffer | Uint8Array | number[]>('mesh_repair_read_positions');
-  let u8: Uint8Array;
-  if (bytes instanceof ArrayBuffer) {
-    u8 = new Uint8Array(bytes);
-  } else if (bytes instanceof Uint8Array) {
-    u8 = bytes;
-  } else if (Array.isArray(bytes)) {
-    u8 = new Uint8Array(bytes);
-  } else {
-    throw new Error('mesh_repair_read_positions returned unexpected type');
-  }
-  // Copy into an aligned ArrayBuffer so the resulting Float32Array is self-contained.
-  const copy = new Uint8Array(u8.byteLength);
-  copy.set(u8);
-  return new Float32Array(copy.buffer);
-}
-
 /**
  * Runs the native mesh-repair engine on a file the Rust side can read
  * directly (STL/OBJ/3MF). Returns null if the current runtime isn't Tauri.
@@ -281,7 +243,7 @@ export async function repairFromPath(
     optionsJson,
   });
   const report = normalizeMeshHealthReport(JSON.parse(reportJson));
-  const positions = await readStagedPositions(core.invoke);
+  const positions = await readPositionsFromCommand(core.invoke, 'mesh_repair_read_positions');
   return { report, positions };
 }
 
@@ -351,7 +313,7 @@ export async function repairFromGeometry(
     optionsJson,
   });
   const report = normalizeMeshHealthReport(JSON.parse(reportJson));
-  const positions = await readStagedPositions(core.invoke);
+  const positions = await readPositionsFromCommand(core.invoke, 'mesh_repair_read_positions');
   return { report, positions };
 }
 
@@ -386,7 +348,7 @@ export async function classifyFromGeometry(
     optionsJson,
   });
   const report = normalizeMeshHealthReport(JSON.parse(reportJson));
-  const positions = await readStagedPositions(core.invoke);
+  const positions = await readPositionsFromCommand(core.invoke, 'mesh_repair_read_positions');
   return { report, positions };
 }
 
@@ -411,27 +373,3 @@ export function applyRepairedPositions(
   geometry.computeBoundingSphere();
 }
 
-function expandGeometryToTriangleSoup(geometry: THREE.BufferGeometry): Float32Array {
-  const posAttr = geometry.getAttribute('position') as THREE.BufferAttribute;
-  const positions = posAttr.array as Float32Array;
-  const index = geometry.getIndex();
-
-  if (!index) {
-    // Already triangle soup, but caller may pass non-Float32Array — coerce.
-    if (positions instanceof Float32Array) {
-      return positions;
-    }
-    return new Float32Array(positions as unknown as ArrayLike<number>);
-  }
-
-  const indexArr = index.array as Uint16Array | Uint32Array;
-  const out = new Float32Array(indexArr.length * 3);
-  for (let i = 0; i < indexArr.length; i++) {
-    const vi = indexArr[i] * 3;
-    const oi = i * 3;
-    out[oi] = positions[vi];
-    out[oi + 1] = positions[vi + 1];
-    out[oi + 2] = positions[vi + 2];
-  }
-  return out;
-}
